services/gps2tsdb/requirements.txt
postgres
pycairo
PyGObject
dbus-python
